backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-docx==1.1.2
pdfminer.six==20231228
PyMuPDF==1.24.10
pytesseract==0.3.13
Pillow==10.4.0
weasyprint==61.2
lxml==5.3.0


